fix(android): :zap: codes that implemented `Service` class are removed
In the Android operating system, we had to implement a service class instance that manages the lifecycle of the module's functions, but it was difficult to manage the process at the time of context injection and termination. Boldly abandoning service class implementation, questioning the need for a background behavior feature. Checking normal behavior.