chore(example): 💄 implementation log display and found new issue about iOS
A log console was created on the example application screen
to debug library deliverables. The speed at which the event is
delivered on Android and the speed at which the event is
delivered on iOS are quite different, so I think we need to
register this part in the issue.