feat(android): configure source sets for new architecture
This commit configures the source sets in the `android/build.gradle` file to include different directories based on whether the new architecture is enabled.

When the new architecture is enabled, the `src/newarch` directory and the generated codegen directory are added to the `java.srcDirs`. Otherwise, the `src/oldarch` directory is added. This allows the application to use different source code depending on the architecture being used.