fix: correct copyright and remove new architecture instructions
This commit corrects a typo in the copyright notice in the LICENSE file and removes outdated instructions for building with the new architecture from the CONTRIBUTING.md file.

The copyright notice was updated to correct a misspelling of the copyright holder's name.

The instructions for building with the new architecture were removed because they are no longer necessary. The new architecture is now enabled by default. Also, the `POD_INSTALL=0` option is no longer needed.